added in the capture functionality to the CC script

+++ UnityProject/Assets/Scripts/Core/Camera/CaptureScreen.cs
@@ -37,7 +37,15 @@ public class CaptureScreen : MonoBehaviour
 			result.ReadPixels(rect, 0, 0);
 
 			byte[] byteArray = result.EncodeToPNG();
-			File.WriteAllBytes(Application.persistentDataPath + Path + "/" + FileName, byteArray);
+			if(Directory.Exists(Application.persistentDataPath + Path))
+			{
+				File.WriteAllBytes(Application.persistentDataPath + Path + "/" + FileName, byteArray);
+			}
+			else
+			{
+				Directory.CreateDirectory(Application.persistentDataPath + Path);
+				File.WriteAllBytes(Application.persistentDataPath + Path + "/" + FileName, byteArray);
+			}
 
 			RenderTexture.ReleaseTemporary(texture);
 			cam.targetTexture = null;
